% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_functions.R
\name{grow_bonsai}
\alias{grow_bonsai}
\title{Grow Bonsai}
\usage{
grow_bonsai(
  seed = 1987,
  time = 6,
  scale = c(0.8, 0.9),
  angle = c(-10, 10),
  angle_z = c(-10, 10),
  split = 2,
  prune = 0,
  shrink_rate = 15
)
}
\arguments{
\item{seed}{Existing sampling}

\item{time}{Total number of layers.}

\item{scale}{Possible values to scale the branch from layer to layer.}

\item{angle}{Possible}

\item{split}{Number of split branches per layer.}

\item{prune}{Probability of pruning a branch.}
}
\description{
Grow the entire tree.
}
\keyword{internal}
