% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_functions.R
\name{grow_layer}
\alias{grow_layer}
\title{Grow Layer}
\usage{
grow_layer(shoots, time, param)
}
\arguments{
\item{shoots}{Number of shoots.}

\item{time}{Branch depth.}

\item{param}{Params.}
}
\description{
To grow a "layer" of the shrub, we extend (and possibly prune) each
existing shoot multiple times
}
\keyword{internal}
