% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_functions.R
\name{grow_shoots}
\alias{grow_shoots}
\title{Grow Shoots}
\usage{
grow_shoots(time, shoots, param)
}
\arguments{
\item{time}{Branch depth.}

\item{shoots}{Number of shoots.}

\item{param}{Parameters defining the tree.}
}
\description{
For each existing shoot on the tree, grow an additional shoot that
extends it; then prune some of them away
}
\keyword{internal}
