% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_functions.R
\name{grow_tree}
\alias{grow_tree}
\title{Grow Tree}
\usage{
grow_tree(sapling, param)
}
\arguments{
\item{sapling}{Existing sampling}

\item{param}{Parameters defining the tree.}
}
\description{
To grow the whole tree we need to "accumulate" the growth: starting with
}
\keyword{internal}
