% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tree.R
\name{render_tree}
\alias{render_tree}
\title{Render Tree}
\usage{
render_tree(
  scene,
  ground_radius = 10,
  ground = TRUE,
  ground_color1 = "darkgreen",
  ground_color2 = "lightgreen",
  fov = NULL,
  lookfrom = NULL,
  lookat = NULL,
  angle = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  lights = TRUE,
  lightintensity = 60,
  clamp_value = 10,
  width = 600,
  height = 600,
  ...
)
}
\arguments{
\item{scene}{Scene of tree model, to be passed to `rayrender`.}

\item{ground_radius}{Default `10`. Radius of the ground.}

\item{ground}{Default `TRUE`. Whether to add a grassy ground scene to the tree.}

\item{ground_color1}{`darkgreen`. Primary ground color.}

\item{ground_color2}{`lightgreen`. Secondary ground color.}

\item{fov}{Default `NULL`, automatically calculated. Camera field of view.}

\item{lookfrom}{Default `NULL`. Camera position. Automatically calculated unless specified.}

\item{lookat}{Default `NULL`. Position camera is directed at. Automatically calculated unless specified.}

\item{angle}{Default `c(0,0,0)`. Degrees to rotate the tree around the X, Y, and Z axes. If this
is a single number, it will be taken as the Y axis rotation.}

\item{order_rotation}{Default `c(1,2,3)`. What order to apply the rotations specified in `angle`.}

\item{lights}{Default `TRUE`. If `FALSE`, removes all default lights.}

\item{lightintensity}{Default `80`. Light intensity.}

\item{clamp_value}{Default `10`. Amount of clamp the light intensity. Finite values help reduce rendering
artifacts, set to `Inf` to turn off this feature.}

\item{width}{Default `600`. Width, in pixels, of the rendered image.}

\item{height}{Default `600`. Height, in pixels, of the rendered image.}

\item{...}{Other arguments to pass to rayrender::render_scene()}
}
\value{
Rayrender scene.
}
\description{
Automatically plots the tree with a camera position and field of view that includes the full model.
For more control over the scene, pass the scene to `rayrender::render_scene()` and specify
the camera position manually. Note: spheres and cylinders in the scene are used to automatically
compute the field of view of the scene--adding additional sphere (e.g. with `rayrender::generate_ground()`)
will change this calculation. Use `rayrender::render_scene()` instead if this is a problem.
}
\examples{
# Generate a basic scene with the default tree.
library(rayrender)
\donttest{
generate_tree() \%>\%
  render_tree()

#Rotate the whole scene
generate_tree() \%>\%
  render_tree(angle = c(0,90,0))

#Specify a custom camera position/direction/field of view/aperture
generate_tree() \%>\%
  render_tree(lookfrom = c(3, 0, 1), lookat = c(0,4,1), fov=30, aperture=1)

#Change the ground color
generate_tree() \%>\%
  render_tree(ground_color1 = "brown", ground_color2 = "orange")

#Turn off lights and add our own
generate_tree() \%>\%
  add_object(sphere(x=20,material=light(color="magenta",intensity=400))) \%>\%
  add_object(sphere(x=-20,material=light(color="lightblue",intensity=400))) \%>\%
  render_tree(lights = FALSE, clamp_value = 10)
}
}
