% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_functions.R
\name{shape_tree}
\alias{shape_tree}
\title{Grow Layer}
\usage{
shape_tree(tree, shrink_rate = 15)
}
\arguments{
\item{tree}{Tree to shape.}
}
\description{
To grow a "layer" of the shrub, we extend (and possibly prune) each
existing shoot multiple times
}
\keyword{internal}
